app/requirements.txt
Django==3.2.6
gunicorn==20.1.0
mysqlclient==2.1.1
django-bootstrap-v5==1.0.11
polygon-api-client==1.4.0
Pillow===9.3.0
django-bootstrap-icons==0.8.2
django-bootstrap-datepicker-plus==5.0.2
python-dateutil